# sitigrid-lambda/sitigrid-API-swagger.yml
...
swagger: "2.0"
info:
  version: "2019-11-13T11:12:00Z"
  title: "Sitigrid API"
host: "gwb6he85bd.execute-api.us-east-1.amazonaws.com"
basePath: "/dev"
schemes:
- "https"
paths:
  /consumptions:
    get:
      description: "Returns all the consumption records in the blockchain"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
      x-amazon-apigateway-integration:
        uri: "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:071760450533:function:sitigrid-fabric-lambda/invocations"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_templates"
        httpMethod: "POST"
        requestTemplates:
          application/json: "{\n  \"functionType\": \"queryObject\",\n  \"chaincodeFunction\"\
            : \"queryAllConsumptions\",\n  \"chaincodeFunctionArgs\": {},\n  \"fabricUsername\"\
            : \"lambdaUser\"\n}"
        contentHandling: "CONVERT_TO_TEXT"
        type: "aws"
    post:
      description: "Adds a new consumption record to the blockchain"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Consumption"
        required: true
        schema:
          $ref: "#/definitions/Consumption"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/TransactionId"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
      x-amazon-apigateway-integration:
        uri: "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:071760450533:function:sitigrid-fabric-lambda/invocations"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_templates"
        httpMethod: "POST"
        requestTemplates:
          application/json: "{\n  \"functionType\": \"invoke\",\n  \"chaincodeFunction\"\
            : \"createConsumptionRecord\",\n  \"chaincodeFunctionArgs\": {\n     \
            \ #set( $body = $util.parseJson($input.body) )\n      \"consumptionId\"\
            : \"$context.requestId\",\n      \"meterpoint\": \"$body.meterpoint\"\
            ,\n      \"consumptionAmount\": $body.consumptionAmount,\n      \"consumptionDate\"\
            :\"$context.requestTimeEpoch\"\n  },\n  \"fabricUsername\": \"lambdaUser\"\
            \n}"
        contentHandling: "CONVERT_TO_TEXT"
        type: "aws"
    options:
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
            Access-Control-Allow-Methods:
              type: "string"
            Access-Control-Allow-Headers:
              type: "string"
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'GET,OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_match"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        type: "mock"
  /meters:
    get:
      description: "Returns all the meters registered"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
      x-amazon-apigateway-integration:
        uri: "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:071760450533:function:sitigrid-fabric-lambda/invocations"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_templates"
        httpMethod: "POST"
        requestTemplates:
          application/json: "{\n  \"functionType\": \"queryObject\",\n  \"chaincodeFunction\"\
            : \"queryAllMeterpoints\",\n  \"chaincodeFunctionArgs\": {},\n  \"fabricUsername\"\
            : \"lambdaUser\"\n}"
        contentHandling: "CONVERT_TO_TEXT"
        type: "aws"
    post:
      description: "Registers a new meterpoint in the blockchain"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Meterpoint"
        required: true
        schema:
          $ref: "#/definitions/Meterpoint"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/TransactionId"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
      x-amazon-apigateway-integration:
        uri: "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:071760450533:function:sitigrid-fabric-lambda/invocations"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_templates"
        httpMethod: "POST"
        requestTemplates:
          application/json: "{\n  \"functionType\": \"invoke\",\n  \"chaincodeFunction\"\
            : \"createMeterpoint\",\n  \"chaincodeFunctionArgs\": {\n      #set( $body\
            \ = $util.parseJson($input.body) )\n      \"meterpoint\": \"$body.meterpoint\"\
            ,\n   \"registeredDate\":\"$context.requestTimeEpoch\"\
            \n  },\n  \"fabricUsername\": \"lambdaUser\"\n}"
        contentHandling: "CONVERT_TO_TEXT"
        type: "aws"
    options:
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
            Access-Control-Allow-Methods:
              type: "string"
            Access-Control-Allow-Headers:
              type: "string"
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'GET,OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_match"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        type: "mock"
  /meters/{meterpoint}:
    get:
      description: "Returns the meterpoint record of the given meterpoint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "meterpoint"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
      x-amazon-apigateway-integration:
        uri: "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:071760450533:function:sitigrid-fabric-lambda/invocations"
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_templates"
        httpMethod: "POST"
        requestTemplates:
          application/json: "{\n  \"functionType\": \"queryObject\",\n  \"chaincodeFunction\"\
            : \"queryMeterpoint\",\n  \"chaincodeFunctionArgs\": {\n      \"meterpoint\"\
            : \"$input.params('meterpoint')\"\n  },\n  \"fabricUsername\": \"lambdaUser\"\
            \n}"
        contentHandling: "CONVERT_TO_TEXT"
        type: "aws"
    options:
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "200 response"
          schema:
            $ref: "#/definitions/Empty"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
            Access-Control-Allow-Methods:
              type: "string"
            Access-Control-Allow-Headers:
              type: "string"
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'GET,OPTIONS'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        passthroughBehavior: "when_no_match"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        type: "mock"
...
definitions:
  Consumption:
    type: "object"
    required:
    - "consumptionAmount"
    - "meterpoint"
    properties:
      consumptionAmount:
        type: "integer"
      meterpoint:
        type: "string"
  Empty:
    type: "object"
    title: "Empty Schema"
  Meterpoint:
    type: "object"
    required:
    - "meterpoint"
    properties:
      meterpoint:
        type: "string"
...
  TransactionId:
    type: "object"
    properties:
      transactionId:
        type: "string"
...